Authentication done via JWT

+++ controllers/api/v1/post_api.js
@@ -19,13 +19,18 @@ module.exports.index = async function (req, res) {
 module.exports.deletePost = async function (req, res) {
   try {
     const post = await Post.findById(req.params.id);
+    if (post.user == req.user.id) {
+      post.remove();
 
-    post.remove();
-
-    await Comment.deleteMany({ post: req.params.id });
-    return res.json(200, {
-      message: "Post and Associated Comment deleted Successfully",
-    });
+      await Comment.deleteMany({ post: req.params.id });
+      return res.json(200, {
+        message: "Post and Associated Comment deleted Successfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post",
+      });
+    }
   } catch (err) {
     return res.json(500, {
       message: "Internal Server Error Solve it first",
